GCP v4 Improvements

+++ gcp/dockerfiles/guide/vacuum.sh
@@ -1,11 +1,9 @@
-#!/bin/bash
+#!/bin/sh
 
-RUN_VACUUM=${RUN_VACUUM:-"no"}
-if [[ "$RUN_VACUUM" != "yes" ]] ; then
-    exit 0
-fi
+[ "${RUN_VACUUM}" == "yes" ] || exit 0
 
-# sleep for a random amount of time, so that we don't run this at the same time on all nodes.
-sleep $[ ( $RANDOM % 240 )  + 1 ]
+DELAY=$(($RANDOM % 240))
+echo Sleep for ${DELAY}s, so that we don\'t run this at the same time on all nodes.
+[ "${SLEEP}" == "no" ] || sleep $(($RANDOM % 240))
 
 docker system prune --force
